merge: merged back header changes, missing newlines and other minor things from linuxbranch
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@10647 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/xbmc/FileSystem/HDHomeRun.cpp
@@ -1,5 +1,5 @@
 #include "stdafx.h"
-#include "../util.h"
+#include "../Util.h"
 
 #include "HDHomeRun.h"
 
@@ -178,7 +178,7 @@ unsigned int CFileHomeRun::Read(void* lpBuf, __int64 uiBufSize)
   DWORD timestamp = GetTickCount() + 5000;
   while(1) 
   {
-    datasize = (unsigned int)min(uiBufSize,UINT_MAX);
+    datasize = (unsigned int)min((unsigned int) uiBufSize,UINT_MAX);
     uint8_t* ptr = m_dll.device_stream_recv(m_device, datasize, &datasize);
     if(ptr)
     {
